+++ client/components/App.tsx
@@ -2,7 +2,7 @@ import React, { useState, FC, useEffect } from "react";
 import { Route, Routes, useNavigate } from "react-router-dom";
 import Login from "./Login";
 import Signup from "./Signup";
-
+import { checkGitHubToken } from "../../client/features/oauth";
 import { ThemeProvider } from "./ClusterView/ClusterViewHeader/themeContext";
 import Dashboard from "./ClusterView/Dashboard";
 
@@ -14,33 +14,7 @@ const App: FC<Props> = () => {
   const navigate = useNavigate();
 
   // effect hook will get access token if
-  useEffect(() => {
-    const queryString = window.location.search;
-    const urlParams = new URLSearchParams(queryString);
-    const codeParam = urlParams.get("code");
-    console.log(codeParam);
-
-    if (codeParam && localStorage.getItem("accessToken") === null) {
-      async function getAccessToken() {
-        await fetch("/user/getAccessToken?code=" + codeParam, {
-          method: "GET",
-        })
-          .then((response) => {
-            return response.json();
-          })
-          .then((data) => {
-            console.log(data);
-
-            if (data.access_token) {
-              localStorage.setItem("accessToken", data.access_token);
-              setRerender(!rerender);
-              navigate("/dashboard");
-            }
-          });
-      }
-      getAccessToken();
-    }
-  }, []);
+  useEffect(() => checkGitHubToken(navigate), []);
 
   return (
     <>
