+++ client/features/oauth/index.ts
@@ -1,4 +1,5 @@
 import { GitHubSignIn } from "./components/GitHubSignIn";
 import { githubLogin } from "./utils/githubLogin";
+import { checkGitHubToken } from "./utils/checkGitHubToken";
 
-export { githubLogin, GitHubSignIn };
+export { githubLogin, GitHubSignIn, checkGitHubToken };
