+++ package.json
@@ -17,7 +17,7 @@
   "dependencies": {
     "@types/react-icons": "^3.0.0",
     "@types/supertest": "^2.0.12",
-    "axios": "^1.4.0",
+    "axios": "^1.5.0",
     "bcrypt": "^5.1.0",
     "bootstrap": "^5.3.0",
     "concurrently": "^8.2.0",
